Ferrocene Self Test: move linker arg parsing
