+++ ferrocene/tools/self-test/Cargo.toml
@@ -9,3 +9,6 @@ edition = "2021"
 [dependencies]
 atty = "0.2.14"
 tempfile = "3.3.0"
+
+[dev-dependencies]
+insta = { version = "1.31.0" }
