+++ ferrocene/tools/self-test/src/linkers/snapshots/rationalise_linker_args.input
@@ -1,0 +1,51 @@
+-plugin
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/liblto_plugin.so
+-plugin-opt=/usr/lib/gcc-cross/aarch64-linux-gnu/12/lto-wrapper
+-plugin-opt=-fresolution=/tmp/cc9GsGIt.res
+-plugin-opt=-pass-through=-lgcc
+-plugin-opt=-pass-through=-lgcc_s
+-plugin-opt=-pass-through=-lc
+-plugin-opt=-pass-through=-lgcc
+-plugin-opt=-pass-through=-lgcc_s
+--sysroot=/
+--build-id
+--eh-frame-hdr
+--hash-style=gnu
+--as-needed
+-dynamic-linker
+/lib/ld-linux-aarch64.so.1
+-X
+-EL
+-maarch64linux
+--fix-cortex-a53-843419
+-pie
+-z
+now
+-z
+relro
+-o
+/tmp/.tmpoNgAJE/output.bin
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/../../../../aarch64-linux-gnu/lib/../lib/Scrt1.o
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/../../../../aarch64-linux-gnu/lib/../lib/crti.o
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/crtbeginS.o
+-L/tmp/.tmpoNgAJE
+-L/usr/lib/gcc-cross/aarch64-linux-gnu/12
+-L/usr/lib/gcc-cross/aarch64-linux-gnu/12/../../../../aarch64-linux-gnu/lib/../lib
+-L/lib/aarch64-linux-gnu
+-L/lib/../lib
+-L/usr/lib/aarch64-linux-gnu
+-L/usr/lib/../lib
+-L/usr/lib/gcc-cross/aarch64-linux-gnu/12/../../../../aarch64-linux-gnu/lib
+/tmp/cc81JpAZ.o
+-lgcc
+--push-state
+--as-needed
+-lgcc_s
+--pop-state
+-lc
+-lgcc
+--as-needed
+-lgcc_s
+--no-as-needed
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/crtendS.o
+/usr/lib/gcc-cross/aarch64-linux-gnu/12/../../../../aarch64-linux-gnu/lib/../lib/crtn.o
