+++ ferrocene/tools/self-test/src/linkers/snapshots/rationalise_linker_args.snap.license
@@ -1,0 +1,2 @@
+SPDX-License-Identifier: MIT OR Apache-2.0
+SPDX-FileCopyrightText: The Ferrocene Developers
