+++ ferrocene/tools/self-test/src/linkers/snapshots/rationalise_linker_extra_args.input
@@ -1,0 +1,5 @@
+--discard-all
+--library-path=/path/to/library
+-I/path/to/dynamic/linker
+-pic-executable
+--no-pie
